requirements.txt
django~=4.1.7
django-extensions~=3.2.1
cryptography~=39.0.2
pyjwt~=2.6.0
whitenoise~=6.4.0
